refactor(config): move to separate package
